# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ClinicFlow AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics/{clinicId}/chat/start:
    post:
      operationId: startChat
      tags: [chat]
      summary: Start a new chat session
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Initial greeting message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /clinics/{clinicId}/chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in the chat flow
      parameters:
        - $ref: "#/components/parameters/ClinicIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: Bot reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
components:
  parameters:
    ClinicIdParam:
      name: clinicId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clinicId:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        patientProblem:
          type: string
        appointmentDate:
          type: string
        sessionId:
          type: ["string", "null"]
        selectedTimeSlot:
          type: ["string", "null"]
        calendarEventId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, pending_slot_selection, confirmed, booked, completed, cancelled]
        appointmentSource:
          type: string
          enum: [Online, Phone, Walk-in, Manual]
        patientAge:
          type: ["integer", "null"]
        patientGender:
          type: ["string", "null"]
        visitType:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        doctorId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, clinicId, patientName, patientPhone, patientProblem, appointmentDate, status, createdAt]
...
    ChatMessageInput:
      type: object
      properties:
        sessionId:
          type: string
        step:
          type: string
        message:
          type: string
          minLength: 1
        context:
          type: object
          properties:
            patientName:
              type: ["string", "null"]
            patientPhone:
              type: ["string", "null"]
            patientProblem:
              type: ["string", "null"]
            appointmentDate:
              type: ["string", "null"]
            selectedTimeSlot:
              type: ["string", "null"]
      required: [sessionId, step, message]
    ChatResponse:
      type: object
      properties:
        sessionId:
          type: string
        step:
          type: string
        botMessage:
          type: string
        isComplete:
          type: boolean
        appointment:
          $ref: "#/components/schemas/Appointment"
      required: [sessionId, step, botMessage, isComplete]
...